+++ app/src/main/kotlin/com/example/bookkeeper/data/model/DmsDigit.kt
@@ -11,12 +11,12 @@ class DmsDigit {
      * 主键
      */
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0L
 
     /**
      * 数字
      */
-    var num: Int? = null
+    var num: Long? = null
 
     /**
      * 生肖
