+++ app/src/main/kotlin/com/example/bookkeeper/data/model/LmsLotteryDetail.kt
@@ -10,23 +10,23 @@ class LmsLotteryDetail {
      * 主键
      */
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0L
 
     /**
      * 彩票ID
      */
     @ColumnInfo(name = "lot_id")
-    var lotId: Int? = null
+    var lotId: Long? = null
 
     /**
      * 数字
      */
-    var num: Int? = null
+    var num: Long? = null
 
     /**
      * 生肖
      */
-    var zod: Int? = null
+    var zod: Long? = null
 
     /**
      * 排序
